fix : gitignore 변경 - #zzam

+++ src/main/java/dongguk/capstone/backend/report/dto/response/ReportResSummaryDTO.java
@@ -16,13 +16,13 @@ public class ReportResSummaryDTO {
     private Long totalConsumption;
 
     @Schema(description = "gpt가 짜준 예산에 비해 몇 %인지 계산한 결과 %")
-    private double percentageOfBudget;
+    private int percentageOfBudget;
 
     @Schema(description = "저번 달의 소비량이 있는지 여부")
     private boolean lastConsumption;
 
     @Schema(description = "저번 달(요청 Month - 1)의 동일한 날짜에 저장된 총 소비량과 비교하여 몇 % 증가/감소했는지에 대한 %")
-    private BigDecimal percentageChange;
+    private int percentageChange;
 
     @Schema(description = "gpt 조언")
     private String gptAdvice;
@@ -31,11 +31,11 @@ public class ReportResSummaryDTO {
     private String highestCategory;
 
     @Schema(description = "요청 Month의 최고 소비 카테고리의 전 달의 동일한 날짜에 저장된 총 소비량과 비교하여 몇 % 증가/감소했는지에 대한 %")
-    private double highestCategoryPercent;
+    private int highestCategoryPercent;
 
     @Schema(description = "요청 Month의 최저 소비 카테고리")
     private String lowestCategory;
 
     @Schema(description = "요청 Month의 최저 소비 카테고리의 전 달의 동일한 날짜에 저장된 총 소비량과 비교하여 몇 % 증가/감소했는지에 대한 %")
-    private double lowestCategoryPercent;
+    private int lowestCategoryPercent;
 }
